Connect jupyter comm from kernel to frontend
